src/apps/worker/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caddie-worker"
version = "0.1.0"
description = "Celery ingestion worker for Caddie"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.4.0",
    "redis>=5.0.8",
    "pydantic>=2.8.2,<3.0",
    "httpx>=0.27.0",
    "openai>=1.51.0",
    "supabase>=2.7.5",
    "pypdf>=4.3.1",
    "python-docx>=1.1.2",
    "markdown>=3.7",
    "python-dotenv>=1.0.1",
    "dateparser>=1.2.0",
    "spacy>=3.7.4"
]

[tool.setuptools.packages.find]
where = ["."]
